+++ src/main/java/com/extendablechattingbe/extendablechattingbe/common/exception/CustomException.java
@@ -1,6 +1,6 @@
 package com.extendablechattingbe.extendablechattingbe.common.exception;
 
-import com.extendablechattingbe.extendablechattingbe.common.ResponseMessages;
+import com.extendablechattingbe.extendablechattingbe.common.CustomMessages;
 import com.extendablechattingbe.extendablechattingbe.common.response.SimpleResponseDTO;
 import lombok.Getter;
 
@@ -8,8 +8,8 @@ import lombok.Getter;
 public class CustomException extends RuntimeException {
     private final SimpleResponseDTO simpleResponseDTO;
 
-    public CustomException(ResponseMessages responseMessages) {
-        super(responseMessages.getMessage());
-        this.simpleResponseDTO = SimpleResponseDTO.of(responseMessages);
+    public CustomException(CustomMessages customMessages) {
+        super(customMessages.getMessage());
+        this.simpleResponseDTO = SimpleResponseDTO.of(customMessages);
     }
 }
