+++ src/main/java/com/extendablechattingbe/extendablechattingbe/common/handler/GlobalExceptionHandler.java
@@ -7,8 +7,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import static com.extendablechattingbe.extendablechattingbe.common.ResponseMessages.BAD_REQUEST_ERROR;
-import static com.extendablechattingbe.extendablechattingbe.common.ResponseMessages.INTERNAL_SERVER_ERROR;
+import static com.extendablechattingbe.extendablechattingbe.common.CustomMessages.BAD_REQUEST_ERROR;
+import static com.extendablechattingbe.extendablechattingbe.common.CustomMessages.INTERNAL_SERVER_ERROR;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler{
