+++ src/main/java/com/extendablechattingbe/extendablechattingbe/common/response/SimpleResponseDTO.java
@@ -1,6 +1,6 @@
 package com.extendablechattingbe.extendablechattingbe.common.response;
 
-import com.extendablechattingbe.extendablechattingbe.common.ResponseMessages;
+import com.extendablechattingbe.extendablechattingbe.common.CustomMessages;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 
@@ -10,19 +10,19 @@ public class SimpleResponseDTO {
     private HttpStatus status;
     private String message;
 
-    private SimpleResponseDTO(ResponseMessages code) {
+    private SimpleResponseDTO(CustomMessages code) {
         this.message = code.getMessage();
         this.status = code.getStatus();
         this.code = code.getCode();
     }
-    private SimpleResponseDTO(ResponseMessages code,String message){
+    private SimpleResponseDTO(CustomMessages code, String message){
         this(code);
         this.message=message;
     }
-    public static SimpleResponseDTO of(ResponseMessages responseMessages) {
-        return new SimpleResponseDTO(responseMessages);
+    public static SimpleResponseDTO of(CustomMessages customMessages) {
+        return new SimpleResponseDTO(customMessages);
     }
-    public static SimpleResponseDTO of(ResponseMessages responseMessages,String message) {
-        return new SimpleResponseDTO(responseMessages,message);
+    public static SimpleResponseDTO of(CustomMessages customMessages, String message) {
+        return new SimpleResponseDTO(customMessages,message);
     }
 }
